refactor(jangar): extract supporting primitives swarm helpers
