+++ docs/jangar/architecture-inventory.md
@@ -14,8 +14,8 @@ This file is generated by `bun run --cwd services/jangar docs:inventory:write`.
 
 | Rank | Module                                                                 |  LOC |
 | ---- | ---------------------------------------------------------------------- | ---: |
-| 1    | `services/jangar/src/server/supporting-primitives-controller.ts`       | 3352 |
-| 2    | `services/jangar/src/server/codex-judge.ts`                            | 2729 |
+| 1    | `services/jangar/src/server/codex-judge.ts`                            | 2729 |
+| 2    | `services/jangar/src/server/supporting-primitives-controller.ts`       | 2479 |
 | 3    | `services/jangar/src/server/atlas-store.ts`                            | 2219 |
 | 4    | `services/jangar/src/server/orchestration-controller.ts`               | 2140 |
 | 5    | `services/jangar/src/server/chat-completion-encoder.ts`                | 2076 |
