+++ services/jangar/src/server/supporting-primitives-controller.ts
@@ -4,13 +4,73 @@ import { resolveSupportingControllerConfig } from '~/server/controller-runtime-c
 import { resolveBooleanFeatureToggle } from '~/server/feature-flags'
 import { createKubeGateway, type KubeGateway } from '~/server/kube-gateway'
 import { startResourceWatch } from '~/server/kube-watch'
-import { assertClusterScopedForWildcard } from '~/server/namespace-scope'
 import { asRecord, asString, readNested } from '~/server/primitives-http'
 import { createKubernetesClient, RESOURCE_MAP } from '~/server/primitives-kube'
 import { resolveSupportingPrimitivesConfig } from '~/server/supporting-primitives-config'
+import { makeName } from '~/server/supporting-primitives-naming'
+import {
+  buildSwarmAgentIdentity,
+  buildSwarmRuntimeParameters,
+  buildSwarmScheduleAnnotations,
+  deriveStageStaggerMinute,
+  normalizeLabelValue,
+  parseStringList,
+  resolveDefaultWorkloadImage,
+  resolveRequirementPriorityScore,
+  resolveScheduleRuntimeInjection,
+  resolveSwarmHulyIntegration,
+  resolveSwarmPersonaForStage,
+  resolveSwarmRunSecrets,
+  sortRequirementSignalsForDispatch,
+  STAGE_CADENCE_KEY,
+  STAGE_LAST_RUN_KEY,
+  STAGE_NAMES,
+  SWARM_AGENT_WORKER_ID_LABEL,
+  SWARM_REQUIREMENT_ANNOTATION_SIGNAL,
+  SWARM_REQUIREMENT_LABEL_ATTEMPT,
+  SWARM_REQUIREMENT_LABEL_CHANNEL,
+  SWARM_REQUIREMENT_LABEL_FROM,
+  SWARM_REQUIREMENT_LABEL_ID,
+  SWARM_REQUIREMENT_LABEL_TO,
+  SWARM_REQUIREMENT_LABEL_TYPE,
+  SWARM_REQUIREMENT_MAX_DISPATCH_PER_RECONCILE,
+  SWARM_REQUIREMENT_MAX_ATTEMPTS,
+  SWARM_REQUIREMENT_SCOPE_FIELD_LIMIT,
+  SWARM_SCHEDULE_ANNOTATION_IDENTITY,
+  type StageName,
+  type StageTargetRef,
+} from '~/server/supporting-primitives-swarm-config'
+import {
+  ACTIVE_PHASES,
+  cadenceToCron,
+  clampUtf8,
+  collectRecentFailureRuns,
+  collectStaleStageSignals,
+  countConsecutiveFailures,
+  countIdempotencyDuplicates,
+  filterRunsAfterTime,
+  getRunTimestamp,
+  isHulyChannel,
+  isIdempotencyDuplicateRun,
+  makeGenerateName,
+  makeRequirementObjective,
+  normalizeParameterMap,
+  parseDurationToMs,
+  parseTimeOrNull,
+  requirementIdForSignal,
+  resolveLatestSuccessfulRunTime,
+  resolveStageApiVersion,
+  resolveStageEvery,
+  resolveStageTargetRef,
+  resolveStageTargetResource,
+  sortByMostRecentRun,
+  stageScheduleName,
+  stringifyUnknown,
+  TERMINAL_FAILURE_PHASES,
+  TERMINAL_SUCCESS_PHASES,
+} from '~/server/supporting-primitives-swarm-analysis'
 import { shouldApplyStatus } from '~/server/status-utils'
 
-const DEFAULT_NAMESPACES = ['agents']
 const DEFAULT_SUPPORTING_CONTROLLER_ENABLED_FLAG_KEY = 'jangar.supporting_controller.enabled'
 const SWARM_CRD_REFRESH_INTERVAL_MS = 15_000
 
@@ -580,44 +640,6 @@ const applyResourceIfChanged = async (
   return kube.apply(resource)
 }
 
-const hashNameSuffix = (value: string) => {
-  let hash = 0
-  for (let i = 0; i < value.length; i += 1) {
-    hash = (hash << 5) - hash + value.charCodeAt(i)
-    hash |= 0
-  }
-  return Math.abs(hash).toString(36).padStart(8, '0').slice(0, 8)
-}
-
-const makeName = (base: string, suffix: string) => {
-  const sanitized = base
-    .toLowerCase()
-    .replace(/[^a-z0-9-]/g, '-')
-    .replace(/-+/g, '-')
-  const maxBaseLength = 45
-  if (sanitized.length <= maxBaseLength) {
-    return `${sanitized}-${suffix}`
-  }
-  const trimmed = sanitized.slice(0, maxBaseLength)
-  return `${trimmed}-${suffix}`
-}
-
-const makeHashedName = (base: string, suffix: string) => {
-  const sanitized = base
-    .toLowerCase()
-    .replace(/[^a-z0-9-]/g, '-')
-    .replace(/-+/g, '-')
-  const maxBaseLength = 63 - 1 - suffix.length
-  if (sanitized.length <= maxBaseLength) {
-    return `${sanitized}-${suffix}`
-  }
-  const hash = hashNameSuffix(base)
-  const separator = '-'
-  const hashBaseLength = Math.max(1, maxBaseLength - hash.length - separator.length)
-  const hashBase = sanitized.slice(0, hashBaseLength)
-  return `${hashBase}-${hash}-${suffix}`
-}
-
 const buildOwnerRefs = (resource: Record<string, unknown>) => {
   const metadata = asRecord(resource.metadata) ?? {}
   const uid = asString(metadata.uid)
@@ -637,649 +659,6 @@ const buildOwnerRefs = (resource: Record<string, unknown>) => {
 
 const resolveNamespace = (resource: Record<string, unknown>) =>
   asString(readNested(resource, ['metadata', 'namespace'])) ?? 'default'
-
-const STAGE_NAMES = ['discover', 'plan', 'implement', 'verify'] as const
-type StageName = (typeof STAGE_NAMES)[number]
-type StageTargetRef = { kind: 'AgentRun' | 'OrchestrationRun'; name: string; namespace: string }
-
-const STAGE_CADENCE_KEY: Record<StageName, string> = {
-  discover: 'discoverEvery',
-  plan: 'planEvery',
-  implement: 'implementEvery',
-  verify: 'verifyEvery',
-}
-
-const STAGE_AGENT_ROLE: Record<StageName, string> = {
-  discover: 'architect',
-  plan: 'architect',
-  implement: 'engineer',
-  verify: 'deployer',
-}
-
-type SwarmPersonaRole = 'architect' | 'engineer' | 'deployer'
-
-const STAGE_PERSONA_ROLE: Record<StageName, SwarmPersonaRole> = {
-  discover: 'architect',
-  plan: 'architect',
-  implement: 'engineer',
-  verify: 'deployer',
-}
-
-const STAGE_LAST_RUN_KEY: Record<StageName, string> = {
-  discover: 'lastDiscoverAt',
-  plan: 'lastPlanAt',
-  implement: 'lastImplementAt',
-  verify: 'lastVerifyAt',
-}
-
-const STAGE_HOURLY_STAGGER_OFFSET: Record<StageName, number> = {
-  discover: 0,
-  plan: 15,
-  implement: 30,
-  verify: 45,
-}
-
-const TERMINAL_SUCCESS_PHASES = new Set(['succeeded', 'success', 'completed'])
-const TERMINAL_FAILURE_PHASES = new Set(['failed', 'error', 'cancelled'])
-const ACTIVE_PHASES = new Set(['pending', 'running', 'inprogress', 'progressing', 'queued'])
-const SWARM_REQUIREMENT_MAX_DISPATCH_PER_RECONCILE = resolveSupportingPrimitivesConfig(
-  process.env,
-).swarmRequirementMaxDispatchPerReconcile
-const SWARM_REQUIREMENT_LABEL_TYPE = 'swarm.proompteng.ai/type'
-const SWARM_REQUIREMENT_LABEL_TO = 'swarm.proompteng.ai/to'
-const SWARM_REQUIREMENT_LABEL_FROM = 'swarm.proompteng.ai/from'
-const SWARM_REQUIREMENT_LABEL_ID = 'swarm.proompteng.ai/requirement-id'
-const SWARM_REQUIREMENT_LABEL_ATTEMPT = 'swarm.proompteng.ai/requirement-attempt'
-const SWARM_REQUIREMENT_LABEL_CHANNEL = 'swarm.proompteng.ai/requirement-channel'
-const SWARM_REQUIREMENT_ANNOTATION_SIGNAL = 'swarm.proompteng.ai/requirement-signal'
-const SWARM_AGENT_WORKER_ID_LABEL = 'swarm.proompteng.ai/worker-id'
-const SWARM_SCHEDULE_ANNOTATION_WORKER_ID = 'swarm.proompteng.ai/worker-id'
-const SWARM_SCHEDULE_ANNOTATION_IDENTITY = 'swarm.proompteng.ai/agent-identity'
-const SWARM_SCHEDULE_ANNOTATION_ROLE = 'swarm.proompteng.ai/agent-role'
-const SWARM_SCHEDULE_ANNOTATION_OWNER_CHANNEL = 'swarm.proompteng.ai/owner-channel'
-const SWARM_SCHEDULE_ANNOTATION_HULY_BASE_URL = 'swarm.proompteng.ai/huly-base-url'
-const SWARM_SCHEDULE_ANNOTATION_HULY_WORKSPACE = 'swarm.proompteng.ai/huly-workspace'
-const SWARM_SCHEDULE_ANNOTATION_HULY_PROJECT = 'swarm.proompteng.ai/huly-project'
-const SWARM_SCHEDULE_ANNOTATION_HULY_SECRET = 'swarm.proompteng.ai/huly-secret'
-const SWARM_SCHEDULE_ANNOTATION_HULY_SKILL_REF = 'swarm.proompteng.ai/huly-skill-ref'
-const SWARM_SCHEDULE_ANNOTATION_HULY_TOKEN_KEY = 'swarm.proompteng.ai/huly-token-key'
-const SWARM_SCHEDULE_ANNOTATION_HULY_EXPECTED_ACTOR_KEY = 'swarm.proompteng.ai/huly-expected-actor-key'
-const SWARM_SCHEDULE_ANNOTATION_HUMAN_NAME = 'swarm.proompteng.ai/human-name'
-const SWARM_REQUIREMENT_SCOPE_FIELD_LIMIT = resolveSupportingPrimitivesConfig(
-  process.env,
-).swarmRequirementMaxPayloadBytes
-const SWARM_DEFAULT_HULY_BASE_URL = resolveSupportingPrimitivesConfig(process.env).swarmDefaultHulyBaseUrl
-const SWARM_DEFAULT_HULY_SKILL_REF = resolveSupportingPrimitivesConfig(process.env).swarmDefaultHulySkillRef
-const SWARM_REQUIREMENT_MAX_ATTEMPTS = resolveSupportingPrimitivesConfig(process.env).swarmRequirementMaxAttempts
-
-const deriveStageStaggerMinute = (swarmName: string, stage: StageName) => {
-  const base = Number.parseInt(hashNameSuffix(swarmName), 36)
-  const swarmOffset = Number.isFinite(base) ? base % 15 : 0
-  return (STAGE_HOURLY_STAGGER_OFFSET[stage] + swarmOffset) % 60
-}
-
-const normalizeRequirementPriority = (value: string | undefined) => {
-  const normalized = (value ?? '').trim().toLowerCase()
-  if (!normalized) return 2
-  if (
-    normalized === 'p0' ||
-    normalized === 'critical' ||
-    normalized === 'urgent' ||
-    normalized === 'blocker' ||
-    normalized === 'highest'
-  ) {
-    return 0
-  }
-  if (normalized === 'p1' || normalized === 'high') {
-    return 1
-  }
-  if (normalized === 'p2' || normalized === 'medium' || normalized === 'normal') {
-    return 2
-  }
-  if (normalized === 'p3' || normalized === 'low') {
-    return 3
-  }
-  return 2
-}
-
-const parseSignalPayloadRecord = (payload: unknown) => {
-  const directRecord = asRecord(payload)
-  if (directRecord) {
-    return directRecord
-  }
-  if (typeof payload === 'string') {
-    try {
-      const parsed = JSON.parse(payload) as unknown
-      return asRecord(parsed)
-    } catch {
-      return null
-    }
-  }
-  return null
-}
-
-const resolveRequirementPriorityScore = (signal: Record<string, unknown>) => {
-  const signalSpec = asRecord(signal.spec) ?? {}
-  const signalMetadata = asRecord(signal.metadata) ?? {}
-  const payloadRecord = parseSignalPayloadRecord(signalSpec.payload)
-  const payloadContextRecord = payloadRecord ? asRecord(payloadRecord.context) : null
-  const candidates = [
-    asString(signalSpec.priority),
-    asString(signalSpec.severity),
-    asString(payloadRecord?.priority),
-    asString(payloadRecord?.severity),
-    asString(payloadContextRecord?.priority),
-    asString(asRecord(signalMetadata.labels)?.priority),
-  ]
-  for (const candidate of candidates) {
-    if (candidate && candidate.trim().length > 0) {
-      return normalizeRequirementPriority(candidate)
-    }
-  }
-  return 2
-}
-
-const sortRequirementSignalsForDispatch = (signals: Record<string, unknown>[]) => {
-  return signals
-    .map((signal) => {
-      const metadata = asRecord(signal.metadata) ?? {}
-      const signalName = asString(metadata.name) ?? ''
-      const createdAt = asString(metadata.creationTimestamp)
-      const createdAtMs = createdAt ? Date.parse(createdAt) : Number.NaN
-      return {
-        signal,
-        signalName,
-        createdAtMs: Number.isFinite(createdAtMs) ? createdAtMs : Number.MAX_SAFE_INTEGER,
-        priority: resolveRequirementPriorityScore(signal),
-      }
-    })
-    .sort((left, right) => {
-      if (left.priority !== right.priority) {
-        return left.priority - right.priority
-      }
-      if (left.createdAtMs !== right.createdAtMs) {
-        return left.createdAtMs - right.createdAtMs
-      }
-      return left.signalName.localeCompare(right.signalName)
-    })
-    .map((entry) => entry.signal)
-}
-
-type SwarmHulyIntegration = {
-  baseUrl: string
-  workspace?: string
-  project?: string
-  secretName: string
-  skillRef: string
-  personas: Partial<Record<SwarmPersonaRole, SwarmPersona>>
-}
-
-type SwarmAgentIdentity = {
-  workerId: string
-  identity: string
-  role: string
-  humanName: string
-  tokenKey: string
-  expectedActorIdKey: string
-}
-
-type SwarmPersona = {
-  role: SwarmPersonaRole
-  humanName: string
-  workerIdentity: string
-  tokenKey: string
-  expectedActorIdKey: string
-}
-
-const DEFAULT_SWARM_PERSONAS: Record<string, { secretName: string; personas: Record<SwarmPersonaRole, SwarmPersona> }> =
-  {
-    'jangar-control-plane': {
-      secretName: 'huly-api-jangar',
-      personas: {
-        architect: {
-          role: 'architect',
-          humanName: 'Victor Chen',
-          workerIdentity: 'victor-chen-jangar-architect',
-          tokenKey: 'HULY_API_TOKEN_VICTOR_CHEN_JANGAR_ARCHITECT',
-          expectedActorIdKey: 'HULY_EXPECTED_ACTOR_ID_VICTOR_CHEN_JANGAR_ARCHITECT',
-        },
-        engineer: {
-          role: 'engineer',
-          humanName: 'Elise Novak',
-          workerIdentity: 'elise-novak-jangar-engineer',
-          tokenKey: 'HULY_API_TOKEN_ELISE_NOVAK_JANGAR_ENGINEER',
-          expectedActorIdKey: 'HULY_EXPECTED_ACTOR_ID_ELISE_NOVAK_JANGAR_ENGINEER',
-        },
-        deployer: {
-          role: 'deployer',
-          humanName: 'Marco Silva',
-          workerIdentity: 'marco-silva-jangar-deployer',
-          tokenKey: 'HULY_API_TOKEN_MARCO_SILVA_JANGAR_DEPLOYER',
-          expectedActorIdKey: 'HULY_EXPECTED_ACTOR_ID_MARCO_SILVA_JANGAR_DEPLOYER',
-        },
-      },
-    },
-    'torghut-quant': {
-      secretName: 'huly-api-torghut',
-      personas: {
-        architect: {
-          role: 'architect',
-          humanName: 'Gideon Park',
-          workerIdentity: 'gideon-park-torghut-architect',
-          tokenKey: 'HULY_API_TOKEN_GIDEON_PARK_TORGHUT_ARCHITECT',
-          expectedActorIdKey: 'HULY_EXPECTED_ACTOR_ID_GIDEON_PARK_TORGHUT_ARCHITECT',
-        },
-        engineer: {
-          role: 'engineer',
-          humanName: 'Naomi Ibarra',
-          workerIdentity: 'naomi-ibarra-torghut-engineer',
-          tokenKey: 'HULY_API_TOKEN_NAOMI_IBARRA_TORGHUT_ENGINEER',
-          expectedActorIdKey: 'HULY_EXPECTED_ACTOR_ID_NAOMI_IBARRA_TORGHUT_ENGINEER',
-        },
-        deployer: {
-          role: 'deployer',
-          humanName: 'Julian Hart',
-          workerIdentity: 'julian-hart-torghut-deployer',
-          tokenKey: 'HULY_API_TOKEN_JULIAN_HART_TORGHUT_DEPLOYER',
-          expectedActorIdKey: 'HULY_EXPECTED_ACTOR_ID_JULIAN_HART_TORGHUT_DEPLOYER',
-        },
-      },
-    },
-  }
-
-const resolveDefaultSwarmPersonaConfig = (swarmName: string, owner: Record<string, unknown>) => {
-  const ownerId = asString(owner.id)?.toLowerCase() ?? ''
-  if (DEFAULT_SWARM_PERSONAS[swarmName]) {
-    return DEFAULT_SWARM_PERSONAS[swarmName]
-  }
-  if (swarmName.toLowerCase().includes('torghut') || ownerId.includes('trading')) {
-    return DEFAULT_SWARM_PERSONAS['torghut-quant']
-  }
-  return DEFAULT_SWARM_PERSONAS['jangar-control-plane']
-}
-
-const normalizeLabelValue = (value: string) => {
-  const normalized = value
-    .toLowerCase()
-    .replace(/[^a-z0-9.-]/g, '-')
-    .replace(/-+/g, '-')
-    .replace(/^-+/, '')
-    .replace(/-+$/, '')
-  if (!normalized) return 'swarm'
-  const trimmed = normalized
-    .slice(0, 63)
-    .replace(/^[.\-]+/, '')
-    .replace(/[.\-]+$/, '')
-  return trimmed || 'swarm'
-}
-
-const resolveDefaultWorkloadImage = () => {
-  const candidate = resolveSupportingPrimitivesConfig(process.env).defaultWorkloadImage ?? ''
-  return candidate.length > 0 ? candidate : null
-}
-
-const parseStringList = (raw: unknown) => {
-  if (!Array.isArray(raw)) return []
-  return raw.map((item) => (typeof item === 'string' ? item.trim() : '')).filter((item) => item.length > 0)
-}
-
-const mergeUniqueStrings = (...values: string[][]) => {
-  const merged: string[] = []
-  const seen = new Set<string>()
-  for (const list of values) {
-    for (const value of list) {
-      if (seen.has(value)) continue
-      seen.add(value)
-      merged.push(value)
-    }
-  }
-  return merged
-}
-
-const GLOBAL_HULY_SECRET = 'huly-api'
-
-const resolveSwarmRunSecrets = (existingSecrets: string[], explicitHulySecret?: string | null) => {
-  const filtered = existingSecrets.filter((secret) => secret !== GLOBAL_HULY_SECRET)
-  const explicit = explicitHulySecret?.trim()
-  if (!explicit) return filtered
-  return mergeUniqueStrings(filtered, [explicit])
-}
-
-const normalizeHulyBaseUrl = (value: string | null | undefined) => {
-  if (!value) return ''
-  const trimmed = value.trim()
-  if (!trimmed) return ''
-  if (trimmed.toLowerCase().startsWith('huly://')) {
-    return SWARM_DEFAULT_HULY_BASE_URL
-  }
-  try {
-    const url = new URL(trimmed)
-    if (!url.hostname.toLowerCase().includes('huly')) return ''
-    // Frontend endpoints do not expose transactor REST routes in-cluster.
-    if (url.hostname.toLowerCase().startsWith('front.')) {
-      url.hostname = `transactor.${url.hostname.slice('front.'.length)}`
-    }
-    const origin = `${url.protocol}//${url.host}`
-    return origin.replace(/\/+$/, '')
-  } catch {
-    return ''
-  }
-}
-
-const resolveSwarmHulyIntegration = (
-  spec: Record<string, unknown>,
-  owner: Record<string, unknown>,
-  swarmName: string,
-): SwarmHulyIntegration => {
-  const integrations = asRecord(spec.integrations) ?? {}
-  const huly = asRecord(integrations.huly) ?? {}
-  const authSecretRef = asRecord(huly.authSecretRef) ?? {}
-  const baseUrl = normalizeHulyBaseUrl(asString(huly.baseUrl)) || SWARM_DEFAULT_HULY_BASE_URL
-  // Avoid implicit global secret fallback; swarm runs should use explicit per-swarm auth secret refs.
-  const secretName = asString(authSecretRef.name)?.trim() ?? ''
-  const workspace = asString(huly.workspace)?.trim() || undefined
-  const project = asString(huly.project)?.trim() || undefined
-  const skillRef = asString(huly.skillRef)?.trim() || SWARM_DEFAULT_HULY_SKILL_REF
-  const rawPersonas = asRecord(huly.personas) ?? {}
-  const personaEntries = Object.entries(rawPersonas)
-    .map(([role, value]) => {
-      if (role !== 'architect' && role !== 'engineer' && role !== 'deployer') return null
-      const record = asRecord(value) ?? {}
-      const humanName = asString(record.humanName)?.trim() ?? ''
-      const workerIdentity = asString(record.workerIdentity)?.trim() ?? ''
-      const tokenKey = asString(record.tokenKey)?.trim() ?? ''
-      const expectedActorIdKey = asString(record.expectedActorIdKey)?.trim() ?? ''
-      if (!humanName || !workerIdentity || !tokenKey || !expectedActorIdKey) {
-        return {
-          role,
-          humanName,
-          workerIdentity,
-          tokenKey,
-          expectedActorIdKey,
-        } satisfies SwarmPersona
-      }
-      return {
-        role,
-        humanName,
-        workerIdentity,
-        tokenKey,
-        expectedActorIdKey,
-      } satisfies SwarmPersona
-    })
-    .filter((persona): persona is SwarmPersona => persona !== null)
-  const explicitPersonas = Object.fromEntries(personaEntries.map((persona) => [persona.role, persona])) as Partial<
-    Record<SwarmPersonaRole, SwarmPersona>
-  >
-  const defaultConfig = resolveDefaultSwarmPersonaConfig(swarmName, owner)
-  const personas =
-    explicitPersonas.architect && explicitPersonas.engineer && explicitPersonas.deployer
-      ? explicitPersonas
-      : (defaultConfig?.personas ?? explicitPersonas)
-  return {
-    baseUrl,
-    workspace,
-    project,
-    secretName: secretName || defaultConfig?.secretName || '',
-    skillRef,
-    personas,
-  }
-}
-
-const resolveSwarmPersonaForStage = (huly: SwarmHulyIntegration, stage: StageName) => {
-  return huly.personas[STAGE_PERSONA_ROLE[stage]]
-}
-
-const buildSwarmAgentIdentity = (input: {
-  swarmName: string
-  stage: StageName
-  persona: SwarmPersona
-  seedSuffix?: string
-}) => {
-  const seed = `${input.swarmName}:${input.stage}:${input.seedSuffix ?? ''}`
-  const hash = hashNameSuffix(seed)
-  const workerId = `worker-${hash.slice(0, 8)}`
-  return {
-    workerId,
-    identity: input.persona.workerIdentity.slice(0, 120),
-    role: input.persona.role,
-    humanName: input.persona.humanName,
-    tokenKey: input.persona.tokenKey,
-    expectedActorIdKey: input.persona.expectedActorIdKey,
-  } satisfies SwarmAgentIdentity
-}
-
-const buildSwarmRuntimeParameters = (input: {
-  ownerChannel: string | null
-  huly: SwarmHulyIntegration
-  identity: SwarmAgentIdentity
-}) => {
-  const parameters: Record<string, string> = {
-    swarmAgentWorkerId: input.identity.workerId,
-    swarmAgentIdentity: input.identity.identity,
-    swarmAgentRole: input.identity.role,
-    swarmHumanName: input.identity.humanName,
-    swarmAgentTokenKey: input.identity.tokenKey,
-    swarmAgentExpectedActorIdKey: input.identity.expectedActorIdKey,
-    hulyApiBaseUrl: input.huly.baseUrl,
-    hulySkillRef: input.huly.skillRef,
-  }
-  if (input.ownerChannel) {
-    parameters.ownerChannel = input.ownerChannel
-  }
-  if (input.huly.workspace) {
-    parameters.hulyWorkspace = input.huly.workspace
-  }
-  if (input.huly.project) {
-    parameters.hulyProject = input.huly.project
-  }
-  return parameters
-}
-
-const buildSwarmScheduleAnnotations = (input: {
-  ownerChannel: string | null
-  huly: SwarmHulyIntegration
-  identity: SwarmAgentIdentity
-}) => {
-  const annotations: Record<string, string> = {
-    [SWARM_SCHEDULE_ANNOTATION_WORKER_ID]: input.identity.workerId,
-    [SWARM_SCHEDULE_ANNOTATION_IDENTITY]: input.identity.identity,
-    [SWARM_SCHEDULE_ANNOTATION_ROLE]: input.identity.role,
-    [SWARM_SCHEDULE_ANNOTATION_HUMAN_NAME]: input.identity.humanName,
-    [SWARM_SCHEDULE_ANNOTATION_HULY_BASE_URL]: input.huly.baseUrl,
-    [SWARM_SCHEDULE_ANNOTATION_HULY_SECRET]: input.huly.secretName,
-    [SWARM_SCHEDULE_ANNOTATION_HULY_SKILL_REF]: input.huly.skillRef,
-    [SWARM_SCHEDULE_ANNOTATION_HULY_TOKEN_KEY]: input.identity.tokenKey,
-    [SWARM_SCHEDULE_ANNOTATION_HULY_EXPECTED_ACTOR_KEY]: input.identity.expectedActorIdKey,
-  }
-  if (input.ownerChannel) {
-    annotations[SWARM_SCHEDULE_ANNOTATION_OWNER_CHANNEL] = input.ownerChannel
-  }
-  if (input.huly.workspace) {
-    annotations[SWARM_SCHEDULE_ANNOTATION_HULY_WORKSPACE] = input.huly.workspace
-  }
-  if (input.huly.project) {
-    annotations[SWARM_SCHEDULE_ANNOTATION_HULY_PROJECT] = input.huly.project
-  }
-  return annotations
-}
-
-const resolveScheduleRuntimeInjection = (schedule: Record<string, unknown>) => {
-  const annotations = asRecord(readNested(schedule, ['metadata', 'annotations'])) ?? {}
-  const ownerChannel = asString(annotations[SWARM_SCHEDULE_ANNOTATION_OWNER_CHANNEL]) ?? null
-  const workerId = asString(annotations[SWARM_SCHEDULE_ANNOTATION_WORKER_ID])
-  const identity = asString(annotations[SWARM_SCHEDULE_ANNOTATION_IDENTITY])
-  const role = asString(annotations[SWARM_SCHEDULE_ANNOTATION_ROLE])
-  const humanName = asString(annotations[SWARM_SCHEDULE_ANNOTATION_HUMAN_NAME])
-  const hulyBaseUrl = asString(annotations[SWARM_SCHEDULE_ANNOTATION_HULY_BASE_URL])
-  const hulyWorkspace = asString(annotations[SWARM_SCHEDULE_ANNOTATION_HULY_WORKSPACE])
-  const hulyProject = asString(annotations[SWARM_SCHEDULE_ANNOTATION_HULY_PROJECT])
-  const hulySecret = asString(annotations[SWARM_SCHEDULE_ANNOTATION_HULY_SECRET])
-  const hulySkillRef = asString(annotations[SWARM_SCHEDULE_ANNOTATION_HULY_SKILL_REF])
-  const hulyTokenKey = asString(annotations[SWARM_SCHEDULE_ANNOTATION_HULY_TOKEN_KEY])
-  const hulyExpectedActorIdKey = asString(annotations[SWARM_SCHEDULE_ANNOTATION_HULY_EXPECTED_ACTOR_KEY])
-
-  const parameters: Record<string, string> = {}
-  if (ownerChannel) parameters.ownerChannel = ownerChannel
-  if (workerId) parameters.swarmAgentWorkerId = workerId
-  if (identity) parameters.swarmAgentIdentity = identity
-  if (role) parameters.swarmAgentRole = role
-  if (humanName) parameters.swarmHumanName = humanName
-  if (hulyBaseUrl) parameters.hulyApiBaseUrl = hulyBaseUrl
-  if (hulyWorkspace) parameters.hulyWorkspace = hulyWorkspace
-  if (hulyProject) parameters.hulyProject = hulyProject
-  if (hulySkillRef) parameters.hulySkillRef = hulySkillRef
-  if (hulyTokenKey) parameters.swarmAgentTokenKey = hulyTokenKey
-  if (hulyExpectedActorIdKey) parameters.swarmAgentExpectedActorIdKey = hulyExpectedActorIdKey
-
-  return {
-    parameters,
-    hulySecret,
-  }
-}
-
-const parseDurationToMs = (raw: string) => {
-  const value = raw.trim().toLowerCase()
-  const match = /^(\d+)\s*(s|sec|secs|second|seconds|m|min|mins|minute|minutes|h|hr|hrs|hour|hours|d|day|days)$/.exec(
-    value,
-  )
-  if (!match) return null
-  const amount = Number(match[1])
-  if (!Number.isFinite(amount) || amount <= 0) return null
-  const unit = match[2]
-  if (unit.startsWith('s')) return amount * 1000
-  if (unit.startsWith('m')) return amount * 60 * 1000
-  if (unit.startsWith('h')) return amount * 60 * 60 * 1000
-  return amount * 24 * 60 * 60 * 1000
-}
-
-const cadenceToCron = (raw: string, options?: { swarmName?: string; stage?: StageName }) => {
-  const durationMs = parseDurationToMs(raw)
-  if (!durationMs) return null
-  const durationMinutes = durationMs / (60 * 1000)
-
-  if (durationMinutes < 1) {
-    return null
-  }
-
-  if (durationMinutes < 60 && Number.isInteger(durationMinutes)) {
-    return `*/${durationMinutes} * * * *`
-  }
-
-  if (durationMinutes % 60 === 0) {
-    const hours = durationMinutes / 60
-    if (hours >= 1 && hours <= 23 && Number.isInteger(hours)) {
-      if (options?.swarmName && options?.stage) {
-        const minute = deriveStageStaggerMinute(options.swarmName, options.stage)
-        if (hours === 1) {
-          return `${minute} * * * *`
-        }
-        return `${minute} */${hours} * * *`
-      }
-      return `0 */${hours} * * *`
-    }
-  }
-
-  if (durationMinutes % (24 * 60) === 0) {
-    const days = durationMinutes / (24 * 60)
-    if (days >= 1 && days <= 28 && Number.isInteger(days)) {
-      return `0 0 */${days} * *`
-    }
-  }
-
-  return null
-}
-
-const collectStaleStageSignals = (
-  stageConfigs: Array<{
-    stage: StageName
-    enabled: boolean
-    every?: string
-  }>,
-  status: Record<string, unknown>,
-  runs: Record<string, unknown>[],
-  nowMs: number,
-  options?: {
-    ignoreBeforeMs?: number | null
-  },
-) => {
-  const ignoreBeforeMs = options?.ignoreBeforeMs ?? null
-  const staleSignals = []
-  for (const stageConfig of stageConfigs) {
-    if (!stageConfig.enabled) continue
-    if (!stageConfig.every) continue
-    const cadenceMs = parseDurationToMs(stageConfig.every)
-    if (cadenceMs === null || cadenceMs <= 0) continue
-    const stageRuns = runs.filter(
-      (run) => asString(readNested(run, ['metadata', 'labels', 'swarm.proompteng.ai/stage'])) === stageConfig.stage,
-    )
-    const latestRunTime = sortByMostRecentRun(stageRuns).at(0)
-    const fromRuns = parseTimeOrNull(getRunTimestamp(latestRunTime ?? {}))
-    const fromStatus = parseTimeOrNull(asString(readNested(status, [STAGE_LAST_RUN_KEY[stageConfig.stage]]) ?? null))
-    const lastRunMs = [fromRuns, fromStatus]
-      .filter((value): value is number => value !== null)
-      .reduce((left, right) => Math.max(left, right), Number.NEGATIVE_INFINITY)
-    if (!Number.isFinite(lastRunMs) || lastRunMs <= 0) {
-      continue
-    }
-    if (ignoreBeforeMs !== null && lastRunMs <= ignoreBeforeMs) {
-      continue
-    }
-    const ageMs = nowMs - lastRunMs
-    const maxAgeMs = 2 * cadenceMs
-    if (ageMs <= maxAgeMs) {
-      continue
-    }
-    staleSignals.push({
-      stage: stageConfig.stage,
-      lastRunAt: new Date(lastRunMs).toISOString(),
-      ageMs,
-      configuredEveryMs: cadenceMs,
-      configuredEvery: stageConfig.every,
-      staleAfterMs: maxAgeMs,
-    })
-  }
-  return staleSignals
-}
-
-const resolveStageEvery = (spec: Record<string, unknown>, stageSpec: Record<string, unknown>, stage: StageName) => {
-  const stageEvery = asString(stageSpec.every)
-  if (stageEvery) return stageEvery
-  const cadence = asRecord(spec.cadence) ?? {}
-  return asString(cadence[STAGE_CADENCE_KEY[stage]])
-}
-
-const resolveStageTargetRef = (stageSpec: Record<string, unknown>, defaultNamespace: string) => {
-  const targetRef = asRecord(stageSpec.targetRef) ?? {}
-  const kind = asString(targetRef.kind)
-  const name = asString(targetRef.name)
-  const namespace = asString(targetRef.namespace) ?? defaultNamespace
-  if (!kind || !name) return null
-  if (kind !== 'AgentRun' && kind !== 'OrchestrationRun') return null
-  return { kind, name, namespace } satisfies StageTargetRef
-}
-
-const resolveStageApiVersion = (kind: string) => {
-  if (kind === 'AgentRun') return 'agents.proompteng.ai/v1alpha1'
-  if (kind === 'OrchestrationRun') return 'orchestration.proompteng.ai/v1alpha1'
-  return ''
-}
-
-const stageScheduleName = (swarmName: string, stage: StageName) => makeHashedName(swarmName, `${stage}-sched`)
-
-const getRunTimestamp = (resource: Record<string, unknown>) => {
-  return (
-    asString(readNested(resource, ['status', 'startedAt'])) ??
-    asString(readNested(resource, ['status', 'finishedAt'])) ??
-    asString(readNested(resource, ['metadata', 'creationTimestamp']))
-  )
-}
-
-const parseTimeOrNull = (value: string | null | undefined) => {
-  if (!value) return null
-  const parsed = Date.parse(value)
-  if (!Number.isFinite(parsed)) return null
-  return parsed
-}
 
 const scheduleSwarmUnfreezeReconcile = (
   kube: ReturnType<typeof createKubernetesClient>,
@@ -1311,271 +690,6 @@ const scheduleSwarmUnfreezeReconcile = (
   }
 
   scheduleChunk(Math.max(0, untilMs - Date.now()))
-}
-
-const sortByMostRecentRun = (resources: Record<string, unknown>[]) => {
-  return [...resources].sort((left, right) => {
-    const leftTs = parseTimeOrNull(getRunTimestamp(left))
-    const rightTs = parseTimeOrNull(getRunTimestamp(right))
-    if (leftTs === null && rightTs === null) return 0
-    if (leftTs === null) return 1
-    if (rightTs === null) return -1
-    return rightTs - leftTs
-  })
-}
-
-const isIdempotencyDuplicateRun = (resource: Record<string, unknown>) => {
-  const rawConditions = readNested(resource, ['status', 'conditions'])
-  const conditions = Array.isArray(rawConditions) ? rawConditions : []
-  if (conditions.length === 0) return false
-  return conditions.some((condition) => {
-    const record = asRecord(condition)
-    if (!record) return false
-    const reason = (asString(record.reason) ?? '').toLowerCase()
-    const type = (asString(record.type) ?? '').toLowerCase()
-    return reason === 'idempotencykeyinuse' || type === 'duplicate'
-  })
-}
-
-const countConsecutiveFailures = (resources: Record<string, unknown>[]) => {
-  const sorted = sortByMostRecentRun(resources)
-  let failures = 0
-  for (const resource of sorted) {
-    if (isIdempotencyDuplicateRun(resource)) continue
-    const phase = (asString(readNested(resource, ['status', 'phase'])) ?? '').toLowerCase()
-    if (!phase) continue
-    if (TERMINAL_FAILURE_PHASES.has(phase)) {
-      failures += 1
-      continue
-    }
-    if (TERMINAL_SUCCESS_PHASES.has(phase)) break
-    if (ACTIVE_PHASES.has(phase)) break
-  }
-  return failures
-}
-
-const countIdempotencyDuplicates = (resources: Record<string, unknown>[]) => {
-  return resources.filter((resource) => isIdempotencyDuplicateRun(resource)).length
-}
-
-const resolveLatestSuccessfulRunTime = (resources: Record<string, unknown>[]) => {
-  const latestSuccessful = sortByMostRecentRun(resources).find((resource) =>
-    TERMINAL_SUCCESS_PHASES.has((asString(readNested(resource, ['status', 'phase'])) ?? '').toLowerCase()),
-  )
-  return latestSuccessful ? parseTimeOrNull(getRunTimestamp(latestSuccessful)) : null
-}
-
-const collectRecentFailureRuns = (resources: Record<string, unknown>[], limit = 5) => {
-  const sorted = sortByMostRecentRun(resources)
-  const failures = sorted
-    .filter((resource) => {
-      if (isIdempotencyDuplicateRun(resource)) return false
-      const phase = (asString(readNested(resource, ['status', 'phase'])) ?? '').toLowerCase()
-      return TERMINAL_FAILURE_PHASES.has(phase)
-    })
-    .slice(0, limit)
-  return failures.map((resource) => {
-    const name = asString(readNested(resource, ['metadata', 'name'])) ?? 'unknown'
-    const namespace = asString(readNested(resource, ['metadata', 'namespace'])) ?? ''
-    const phase = (asString(readNested(resource, ['status', 'phase'])) ?? '').toLowerCase()
-    const timestamp = getRunTimestamp(resource)
-    const conclusion = asString(readNested(resource, ['status', 'conclusion']))
-    const failedCondition = (
-      Array.isArray(readNested(resource, ['status', 'conditions']))
-        ? (readNested(resource, ['status', 'conditions']) as unknown[])
-        : []
-    )
-      .map((entry) => asRecord(entry))
-      .find((condition) => asString(condition?.type) === 'Failed' && asString(condition?.status) === 'True')
-    const failedStep = (
-      Array.isArray(readNested(resource, ['status', 'workflow', 'steps']))
-        ? (readNested(resource, ['status', 'workflow', 'steps']) as unknown[])
-        : []
-    )
-      .map((entry) => asRecord(entry))
-      .find((step) => asString(step?.phase) === 'Failed')
-    const reason =
-      asString(readNested(resource, ['status', 'message'])) ??
-      asString(readNested(resource, ['status', 'reason'])) ??
-      asString(failedCondition?.message) ??
-      asString(failedCondition?.reason) ??
-      asString(failedStep?.message)
-    return {
-      name,
-      namespace,
-      phase,
-      timestamp,
-      conclusion: conclusion || null,
-      reason: reason || null,
-    }
-  })
-}
-
-const filterRunsAfterTime = (resources: Record<string, unknown>[], timestampMs: number | null) => {
-  if (timestampMs === null) return resources
-  return resources.filter((resource) => {
-    const runTimestamp = parseTimeOrNull(getRunTimestamp(resource))
-    return runTimestamp !== null && runTimestamp > timestampMs
-  })
-}
-
-const requirementIdForSignal = (signalNamespace: string, signalName: string) =>
-  hashNameSuffix(`${signalNamespace}/${signalName}`)
-
-const isHulyChannel = (channel: string | null | undefined) => {
-  if (!channel) return false
-  const value = channel.trim()
-  if (!value) return false
-  if (value.toLowerCase().startsWith('huly://')) return true
-  try {
-    const url = new URL(value)
-    return url.hostname.toLowerCase().includes('huly')
-  } catch {
-    return false
-  }
-}
-
-const stringifyUnknown = (value: unknown) => {
-  if (typeof value === 'string') return value
-  try {
-    const result = JSON.stringify(value)
-    return result === undefined ? '' : result
-  } catch {
-    return ''
-  }
-}
-
-const truncateUtf8 = (value: string, maxBytes: number) => {
-  const safeMaxBytes = Math.max(0, maxBytes)
-  if (!value) return ''
-  if (safeMaxBytes === 0) return ''
-
-  let usedBytes = 0
-  const encoder = new TextEncoder()
-  let result = ''
-
-  for (const char of value) {
-    const charBytes = encoder.encode(char).byteLength
-    if (usedBytes + charBytes > safeMaxBytes) break
-    usedBytes += charBytes
-    result += char
-  }
-
-  return result
-}
-
-const clampUtf8 = (value: string, maxBytes: number) => {
-  const encoder = new TextEncoder()
-  const encoded = encoder.encode(value)
-  if (encoded.byteLength <= maxBytes) {
-    return {
-      value,
-      bytes: encoded.byteLength,
-      truncated: false,
-    }
-  }
-
-  return {
-    value: truncateUtf8(value, maxBytes),
-    bytes: encoded.byteLength,
-    truncated: true,
-  }
-}
-
-const extractPayloadObjective = (payload: string) => {
-  try {
-    const parsed = JSON.parse(payload)
-    if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) {
-      return ''
-    }
-
-    const objectiveValue = parsed.objective
-    if (typeof objectiveValue === 'string') {
-      const normalized = objectiveValue.trim()
-      if (normalized.length > 0) return normalized
-    }
-
-    const acceptanceValue = parsed.acceptance
-    if (typeof acceptanceValue === 'string') {
-      const normalized = acceptanceValue.trim()
-      if (normalized.length > 0) return `acceptance: ${normalized}`
-    }
-    if (Array.isArray(acceptanceValue)) {
-      const acceptanceList = acceptanceValue
-        .map((value) => (typeof value === 'string' ? value.trim() : ''))
-        .filter((value) => value.length > 0)
-      if (acceptanceList.length > 0) {
-        return `acceptance: ${acceptanceList.join(', ')}`
-      }
-    }
-
-    const missionValue = parsed.mission
-    if (typeof missionValue === 'string') {
-      const normalized = missionValue.trim()
-      if (normalized.length > 0) return `mission: ${normalized}`
-    }
-
-    const objectiveSource = extractPayloadObjectiveObject(parsed as Record<string, unknown>)
-    if (objectiveSource) {
-      return objectiveSource
-    }
-  } catch {
-    // ignore malformed payloads
-  }
-  return ''
-}
-
-const extractPayloadObjectiveObject = (payload: Record<string, unknown>) => {
-  const fields: Array<unknown> = [payload.objective, payload.scope, payload.goal, payload.summary]
-  for (const value of fields) {
-    if (typeof value === 'string') {
-      const normalized = value.trim()
-      if (normalized.length > 0) return normalized
-    }
-    if (typeof value === 'number' || typeof value === 'boolean') {
-      return String(value)
-    }
-  }
-  return ''
-}
-
-const makeRequirementObjective = (description: string, payload: string) => {
-  const payloadObjective = payload ? extractPayloadObjective(payload) : ''
-  const objectiveSource = payloadObjective || payload
-  const parts = [description, objectiveSource].filter((value) => value.length > 0)
-  if (parts.length === 0) return ''
-  const combined = parts.join('\n\n')
-  return clampUtf8(combined, SWARM_REQUIREMENT_SCOPE_FIELD_LIMIT).value
-}
-
-const normalizeParameterMap = (raw: unknown) => {
-  const record = asRecord(raw) ?? {}
-  const output: Record<string, string> = {}
-  for (const [key, value] of Object.entries(record)) {
-    if (typeof value === 'string') {
-      output[key] = value
-      continue
-    }
-    if (value === null || value === undefined) continue
-    output[key] = stringifyUnknown(value)
-  }
-  return output
-}
-
-const makeGenerateName = (base: string, suffix: string) => {
-  const candidate = makeHashedName(base, suffix)
-  const trimmed = candidate.slice(0, 62).replace(/-+$/, '')
-  return `${trimmed || 'run'}-`
-}
-
-const resolveStageTargetResource = async (
-  kube: ReturnType<typeof createKubernetesClient>,
-  targetRef: StageTargetRef,
-) => {
-  if (targetRef.kind === 'AgentRun') {
-    return kube.get(RESOURCE_MAP.AgentRun, targetRef.name, targetRef.namespace)
-  }
-  return kube.get(RESOURCE_MAP.OrchestrationRun, targetRef.name, targetRef.namespace)
 }
 
 const reconcileTool = async (kube: ReturnType<typeof createKubernetesClient>, tool: Record<string, unknown>) => {
